refactor(social): split SocialFollowPopup's 5 bundled components into files
SocialFollowPopup.tsx (851 lines) already contained 5 separate,
self-contained top-level components (CloseIcon, IosInstallSteps,
InstallSection, PushSection, SocialLinksSection) defined inline -
pure code motion into their own files, no behavior change. 851 -> 528
lines on the main file.

Follow-up from the oversized-files finding flagged in the previous
audit PR: unlike NavigationFiltersModal (no clean seams) or
publica/page.tsx (needs a real extraction), this one just needed the
existing seams moved out.
