+++ components/ui/common/social/SocialFollowPopup.tsx
@@ -6,9 +6,11 @@ import useCheckMobileScreen from "@components/hooks/useCheckMobileScreen";
 import { usePushNotifications } from "@components/hooks/usePushNotifications";
 import { usePwaInstall } from "@components/hooks/usePwaInstall";
 import { sendGoogleEvent } from "@utils/analytics";
-import { SocialIcon } from "./icons";
 import type { SocialPopupState } from "types/props";
-import { socialLinks } from "@config/index";
+import CloseIcon from "./CloseIcon";
+import InstallSection from "./InstallSection";
+import PushSection from "./PushSection";
+import SocialLinksSection from "./SocialLinksSection";
 
 export const STORAGE_KEY = "social-follow-popup";
 export const PAGE_VIEW_KEY = "social-popup-views";
@@ -74,331 +76,6 @@ function getScrollThreshold(): number {
   );
   const viewportHeight = window.innerHeight;
   return (docHeight - viewportHeight) * SCROLL_PERCENT;
-}
-
-const SOCIAL_LINKS = [
-  { platform: "instagram", label: "Instagram", href: socialLinks.instagram },
-  { platform: "twitter", label: "X", href: socialLinks.twitter },
-  { platform: "facebook", label: "Facebook", href: socialLinks.facebook },
-  { platform: "threads", label: "Threads", href: socialLinks.threads },
-  { platform: "linkedin", label: "LinkedIn", href: socialLinks.linkedin },
-  { platform: "telegram", label: "Telegram", href: socialLinks.telegram },
-  // { platform: "tiktok", label: "TikTok", href: socialLinks.tiktok },
-  { platform: "mastodon", label: "Mastodon", href: socialLinks.mastodon },
-] as const;
-
-/* ------------------------------------------------------------------ */
-/*  Shared sections (single source of truth for mobile + desktop)      */
-/* ------------------------------------------------------------------ */
-
-function CloseIcon() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth={2}
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className="w-5 h-5"
-      aria-hidden="true"
-    >
-      <line x1="18" y1="6" x2="6" y2="18" />
-      <line x1="6" y1="6" x2="18" y2="18" />
-    </svg>
-  );
-}
-
-/**
- * iOS "Add to Home Screen" instruction sheet, following the de-facto
- * standard pattern (react-ios-pwa-prompt, Progressier): app identity row
- * matching the native A2HS preview, then numbered steps with the real iOS
- * glyphs so users pattern-match icons instead of translating words.
- */
-function IosInstallSteps({
-  isIpad,
-  shareLocation,
-}: {
-  isIpad: boolean;
-  shareLocation: "safari" | "menu";
-}) {
-  const t = useTranslations("Components.SocialFollowPopup");
-  const shareStep =
-    shareLocation === "menu"
-      ? t("installStepShareMenu")
-      : isIpad
-        ? t("installStepShareIpad")
-        : t("installStepShare");
-  return (
-    <div className="flex flex-col">
-      {/* Identity row: mirrors what iOS shows in the A2HS dialog */}
-      <div className="flex items-center gap-3 pb-3">
-        {/* eslint-disable-next-line @next/next/no-img-element -- 40px static
-            asset already precached by the SW; next/image adds no value */}
-        <img
-          src="/static/icons/icon-192x192.png"
-          alt=""
-          aria-hidden="true"
-          width={40}
-          height={40}
-          className="rounded-lg border border-border/60 bg-background"
-        />
-        <div className="text-left">
-          <p
-            className="body-small font-semibold text-foreground-strong leading-tight"
-            translate="no"
-          >
-            Esdeveniments
-          </p>
-          <p className="body-small text-foreground/60 leading-tight" translate="no">
-            esdeveniments.cat
-          </p>
-        </div>
-      </div>
-
-      <ol className="flex flex-col" aria-label={t("installEnable")}>
-        <li className="flex items-center gap-3 py-2.5 border-t border-border/40">
-          <span className="body-small text-foreground/50 w-4 text-center flex-shrink-0">
-            1
-          </span>
-          {shareLocation === "menu" ? (
-            /* ⋯ overflow glyph for Chrome/Firefox/Edge/Opera on iOS, where
-               Share sits inside the menu rather than the toolbar */
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="currentColor"
-              className="w-5 h-5 flex-shrink-0 text-foreground"
-              aria-hidden="true"
-            >
-              <circle cx="5" cy="12" r="1.6" />
-              <circle cx="12" cy="12" r="1.6" />
-              <circle cx="19" cy="12" r="1.6" />
-            </svg>
-          ) : (
-            /* iOS share glyph; #007AFF replicates the system tint so users
-               can pattern-match the real button (semantic tokens would
-               defeat the recognition purpose) */
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="#007AFF"
-              strokeWidth={1.8}
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="w-5 h-5 flex-shrink-0"
-              aria-hidden="true"
-            >
-              <path d="M12 3v12M8 7l4-4 4 4" />
-              <path d="M5 11v8a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2v-8" />
-            </svg>
-          )}
-          <span className="body-small text-foreground text-left">
-            {shareStep}
-          </span>
-        </li>
-        <li className="flex items-center gap-3 py-2.5 border-t border-border/40">
-          <span className="body-small text-foreground/50 w-4 text-center flex-shrink-0">
-            2
-          </span>
-          {/* iOS "Add to Home Screen" plus-square glyph */}
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth={1.8}
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            className="w-5 h-5 flex-shrink-0 text-foreground"
-            aria-hidden="true"
-          >
-            <rect x="4" y="4" width="16" height="16" rx="4" />
-            <path d="M12 8.5v7M8.5 12h7" />
-          </svg>
-          <span className="body-small text-foreground text-left">
-            {t("installStepAdd")}
-          </span>
-        </li>
-      </ol>
-    </div>
-  );
-}
-
-function InstallSection({
-  canPromptInstall,
-  showIosInstructions,
-  showOpenInSafariHint,
-  isIpad,
-  iosShareLocation,
-  isInstalling,
-  onInstall,
-}: {
-  canPromptInstall: boolean;
-  showIosInstructions: boolean;
-  showOpenInSafariHint: boolean;
-  isIpad: boolean;
-  iosShareLocation: "safari" | "menu";
-  isInstalling: boolean;
-  onInstall: () => void;
-}) {
-  const t = useTranslations("Components.SocialFollowPopup");
-  if (!canPromptInstall && !showIosInstructions && !showOpenInSafariHint) {
-    return null;
-  }
-  return (
-    <div className="flex flex-col gap-2.5">
-      {canPromptInstall ? (
-        <button
-          onClick={onInstall}
-          disabled={isInstalling}
-          className="btn-primary w-full justify-center"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth={2}
-            className="w-4 h-4"
-            aria-hidden="true"
-          >
-            <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4M7 10l5 5 5-5M12 15V3" />
-          </svg>
-          {isInstalling ? t("installEnabling") : t("installEnable")}
-        </button>
-      ) : null}
-      {showIosInstructions ? (
-        <IosInstallSteps isIpad={isIpad} shareLocation={iosShareLocation} />
-      ) : null}
-      {showOpenInSafariHint ? (
-        <p className="body-small text-foreground/70 text-center leading-relaxed">
-          {t("installInAppHelp")}
-        </p>
-      ) : null}
-    </div>
-  );
-}
-
-function PushSection({
-  pushState,
-  isSubscribing,
-  onSubscribe,
-}: {
-  pushState: ReturnType<typeof usePushNotifications>["state"];
-  isSubscribing: boolean;
-  onSubscribe: () => void;
-}) {
-  const t = useTranslations("Components.SocialFollowPopup");
-  return (
-    <div className="flex flex-col gap-2.5 bg-muted/60 border border-border/40 rounded-card p-4">
-      {/* Persistent live region: must exist in the DOM before its content
-          changes for screen readers to announce it. sr-only (position:absolute)
-          so it adds no layout; the rows below render the same state visually. */}
-      <div role="status" aria-live="polite" className="sr-only">
-        {pushState === "subscribed"
-          ? t("pushEnabled")
-          : pushState === "denied"
-            ? t("pushBlockedHelp")
-            : ""}
-      </div>
-      {pushState === "unsubscribed" ? (
-        <button
-          onClick={onSubscribe}
-          disabled={isSubscribing}
-          className="btn-primary w-full justify-center"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth={2}
-            className="w-4 h-4"
-            aria-hidden="true"
-          >
-            <path d="M18 8h1a4 4 0 0 1 4 4v10a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V12a4 4 0 0 1 4-4h1V6a4 4 0 0 1 8 0v2zM10 19v-6m4 6v-6" />
-          </svg>
-          {isSubscribing ? t("pushEnabling") : t("pushEnable")}
-        </button>
-      ) : null}
-      {pushState === "denied" ? (
-        <p className="body-small text-foreground/70 text-center">
-          {t("pushBlockedHelp")}
-        </p>
-      ) : null}
-      {pushState === "subscribed" ? (
-        <div className="flex items-center justify-center gap-2 py-1">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 24 24"
-            fill="currentColor"
-            className="w-5 h-5 text-success"
-            aria-hidden="true"
-          >
-            <path d="M9 16.17L4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41z" />
-          </svg>
-          <p className="body-small text-success font-medium">
-            {t("pushEnabled")}
-          </p>
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function SocialLinksSection({ variant }: { variant: "mobile" | "desktop" }) {
-  const t = useTranslations("Components.SocialFollowPopup");
-  return (
-    <div className="flex flex-col gap-2.5">
-      <p className="body-small font-semibold text-foreground/60 text-center uppercase tracking-wide">
-        {t("followLabel")}
-      </p>
-      {variant === "mobile" ? (
-        /* Horizontal scroll row. The inner `w-max mx-auto` wrapper centers
-           the chips when they fit and keeps the row left-anchored (fully
-           scrollable) when they overflow — `justify-center` on an
-           overflowing flex container clips its start edge unreachably. */
-        <div className="overflow-x-auto no-scrollbar -mx-1 px-1">
-          <div className="flex w-max mx-auto gap-1.5">
-            {SOCIAL_LINKS.map((social) => (
-              <a
-                key={social.platform}
-                href={social.href}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="flex-center gap-1.5 px-3.5 py-2 rounded-full border border-border/60 bg-muted/30 hover:bg-primary/10 hover:border-primary/40 hover:scale-105 transition-[transform,background-color,border-color] duration-normal text-foreground body-small font-medium no-underline flex-shrink-0 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-offset-2"
-                aria-label={social.label}
-              >
-                <span className="text-primary">
-                  <SocialIcon platform={social.platform} className="w-4 h-4" />
-                </span>
-              </a>
-            ))}
-          </div>
-        </div>
-      ) : (
-        <div className="grid grid-cols-3 gap-2">
-          {SOCIAL_LINKS.map((social) => (
-            <a
-              key={social.platform}
-              href={social.href}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="flex-center gap-1.5 px-3 py-2.5 rounded-card border border-border/60 bg-muted/30 hover:bg-primary/10 hover:border-primary/40 hover:scale-105 transition-[transform,background-color,border-color] duration-normal text-foreground body-small font-medium no-underline focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-offset-2"
-              aria-label={social.label}
-            >
-              <span className="text-primary">
-                <SocialIcon platform={social.platform} className="w-5 h-5" />
-              </span>
-            </a>
-          ))}
-        </div>
-      )}
-    </div>
-  );
 }
 
 export default function SocialFollowPopup({ pathname }: { pathname: string }) {
